Handle multiple BA stats file versions
