+++ shared/bundle_analysis/parsers/__init__.py
@@ -1,0 +1,7 @@
+from shared.bundle_analysis.parsers.v1 import ParserV1
+from shared.bundle_analysis.parsers.v2 import ParserV2
+
+__all__ = [
+    "ParserV1",
+    "ParserV2",
+]
