+++ tests/unit/bundle_analysis/test_bundle_analysis.py
@@ -22,6 +22,10 @@ sample_bundle_stats_path_3 = (
     / "sample_bundle_stats_invalid_name.json"
 )
 
+sample_bundle_stats_path_4 = (
+    Path(__file__).parent.parent.parent / "samples" / "sample_bundle_stats_v1.json"
+)
+
 
 def test_create_bundle_report():
     try:
@@ -43,13 +47,14 @@ def test_create_bundle_report():
         asset_reports = list(bundle_report.asset_reports())
 
         assert [
-            (ar.name, ar.hashed_name, ar.size, len(ar.modules()), ar.asset_type)
+            (ar.name, ar.hashed_name, ar.size, ar.gzip_size, len(ar.modules()), ar.asset_type)
             for ar in asset_reports
         ] == [
             (
                 "assets/react-*.svg",
                 "assets/react-35ef61ed.svg",
                 4126,
+                4125,
                 0,
                 AssetType.IMAGE,
             ),
@@ -57,6 +62,7 @@ def test_create_bundle_report():
                 "assets/index-*.css",
                 "assets/index-d526a0c5.css",
                 1421,
+                1420,
                 0,
                 AssetType.STYLESHEET,
             ),
@@ -64,6 +70,7 @@ def test_create_bundle_report():
                 "assets/LazyComponent-*.js",
                 "assets/LazyComponent-fcbb0922.js",
                 294,
+                293,
                 1,
                 AssetType.JAVASCRIPT,
             ),
@@ -71,6 +78,7 @@ def test_create_bundle_report():
                 "assets/index-*.js",
                 "assets/index-c8676264.js",
                 154,
+                153,
                 2,
                 AssetType.JAVASCRIPT,
             ),
@@ -80,6 +88,7 @@ def test_create_bundle_report():
                 "assets/index-*.js",
                 "assets/index-666d2e09.js",
                 144577,
+                144576,
                 28,
                 AssetType.JAVASCRIPT,
             ),
@@ -251,7 +260,7 @@ def test_bundle_report_info():
     bundle_report = report.bundle_report("sample")
     bundle_report_info = bundle_report.info()
 
-    assert bundle_report_info["version"] == "1"
+    assert bundle_report_info["version"] == "2"
     assert bundle_report_info["bundler_name"] == "rollup"
     assert bundle_report_info["bundler_version"] == "3.29.4"
     assert bundle_report_info["built_at"] == 1701451048604
@@ -275,7 +284,7 @@ def test_bundle_report_size_integer():
 
 def test_bundle_parser_error():
     with patch(
-        "shared.bundle_analysis.parser.Parser._parse_assets_event",
+        "shared.bundle_analysis.parsers.ParserV1._parse_assets_event",
         side_effect=Exception("MockError"),
     ):
         report = BundleAnalysisReport()
@@ -334,3 +343,84 @@ def test_bundle_file_save_unknown_error():
 
             assert str(excinfo) == "UnknownError"
             assert type(excinfo) == Exception
+
+
+def test_create_bundle_report_v1():
+    try:
+        report = BundleAnalysisReport()
+        session_id = report.ingest(sample_bundle_stats_path_4)
+        assert session_id == 1
+
+        assert report.metadata() == {
+            MetadataKey.SCHEMA_VERSION: 2,
+        }
+
+        bundle_reports = list(report.bundle_reports())
+        assert len(bundle_reports) == 1
+
+        bundle_report = report.bundle_report("invalid")
+        assert bundle_report is None
+        bundle_report = report.bundle_report("sample")
+
+        bundle_report_info = bundle_report.info()
+        assert bundle_report_info["version"] == "1"
+
+        asset_reports = list(bundle_report.asset_reports())
+
+        assert [
+            (ar.name, ar.hashed_name, ar.size, ar.gzip_size, len(ar.modules()), ar.asset_type)
+            for ar in asset_reports
+        ] == [
+            (
+                "assets/react-*.svg",
+                "assets/react-35ef61ed.svg",
+                4126,
+                4,
+                0,
+                AssetType.IMAGE,
+            ),
+            (
+                "assets/index-*.css",
+                "assets/index-d526a0c5.css",
+                1421,
+                1,
+                0,
+                AssetType.STYLESHEET,
+            ),
+            (
+                "assets/LazyComponent-*.js",
+                "assets/LazyComponent-fcbb0922.js",
+                294,
+                0,
+                1,
+                AssetType.JAVASCRIPT,
+            ),
+            (
+                "assets/index-*.js",
+                "assets/index-c8676264.js",
+                154,
+                0,
+                2,
+                AssetType.JAVASCRIPT,
+            ),
+            # FIXME: this is wrong since it's capturing the SVG and CSS modules as well.
+            # Made a similar note in the parser code where the associations are made
+            (
+                "assets/index-*.js",
+                "assets/index-666d2e09.js",
+                144577,
+                144,
+                28,
+                AssetType.JAVASCRIPT,
+            ),
+        ]
+
+        for ar in asset_reports:
+            for module in ar.modules():
+                assert isinstance(module.name, str)
+                assert isinstance(module.size, int)
+
+        assert bundle_report.total_size() == 150572
+        assert report.session_count() == 1
+    finally:
+        report.cleanup()
